Move dev/components and dev/colors into their own dir
