+++ elixir/assets/js/hooks.js
@@ -453,7 +453,7 @@ Hooks.ProgressBar = {
 };
 
 if (__DEV__) {
-  Object.assign(Hooks, require("./hooks/dev").default);
+  Object.assign(Hooks, require("../../portal_dev/js/dev").default);
 }
 
 export default Hooks;
