+++ elixir/portal_dev/live/dev/colors_live.ex
@@ -1,5 +1,5 @@
 defmodule PortalWeb.Dev.ColorsLive do
-  use PortalWeb, {:live_view, layout: {PortalWeb.Layouts, :components_preview}}
+  use PortalWeb, {:live_view, layout: {PortalWeb.Dev.Layouts, :components_preview}}
 
   @sections [
     %{
